Uppdaterad SignalR för teams-sidan

+++ server/Controllers/TeamUsersController.cs
@@ -129,21 +129,23 @@ namespace server.Controllers
         //    return StatusCode(StatusCodes.Status201Created, new { teamUserId = teamUser.TeamID, username = user.Username });
         //}
 
-        [HttpDelete("{teamId}/{userId}")]
-        public IActionResult RemoveUserFromTeam(int teamId, int userId)
+        [HttpDelete]
+        public async Task <IActionResult> RemoveUserFromTeam(DropUserFromTeamRequest request)
         {
-            var teamUser = new TeamUser
+            var teamUser = dbContext.TeamUsers
+                .FirstOrDefault(tu => tu.TeamID == request.TeamID && tu.UserId == request.UserID);
+
+            if (teamUser == null)
             {
-                TeamID = teamId,
-                UserId = userId
-            };
-            dbContext.TeamUsers.Attach(teamUser); 
-            dbContext.TeamUsers.Remove(teamUser);   
-            dbContext.SaveChanges();                
+                return NotFound(new { message = "Användaren finns inte i laget." });
+            }
+
+            dbContext.TeamUsers.Remove(teamUser);
+            dbContext.SaveChanges();
+
+            await _hubContext.Clients.All.SendAsync("ReceiveUserLeftTeam", teamUser.TeamID, teamUser.UserId);
+
             return Ok(new { message = "Användaren har tagits bort från laget." });
         }
-
-
-
     }
 }
